# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"API Recetas - Documentación dinámica","description":"Documentación generada dinámicamente de los endpoints activos","contact":{"name":"Equipo Recetas","email":"devops@example.com"},"license":{"name":"MIT"},"version":"v1"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/recetas/{id}":{"get":{"tags":["receta-controller"],"operationId":"getRecetaById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["receta-controller"],"summary":"Actualizar receta (requiere autenticación)","operationId":"updateReceta","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Receta"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["receta-controller"],"summary":"Eliminar receta (requiere autenticación)","operationId":"deleteReceta","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/paises/{id}":{"get":{"tags":["pais-controller"],"operationId":"getPaisById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["pais-controller"],"summary":"Actualizar país (requiere autenticación)","operationId":"updatePais","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Pais"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["pais-controller"],"summary":"Eliminar país (requiere autenticación)","operationId":"deletePais","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/ingredientes/{id}":{"get":{"tags":["ingrediente-controller"],"summary":"Obtener ingrediente por id (público)","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["ingrediente-controller"],"summary":"Actualizar ingrediente (requiere autenticación)","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ingrediente"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["ingrediente-controller"],"summary":"Eliminar ingrediente (requiere autenticación)","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/estrellas/{id}":{"put":{"tags":["estrella-controller"],"summary":"Actualizar estrella (requiere autenticación)","operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["estrella-controller"],"summary":"Eliminar estrella (requiere autenticación)","operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/comentarios/{id}":{"put":{"tags":["comentario-controller"],"summary":"Actualizar comentario (requiere autenticación)","operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["comentario-controller"],"summary":"Eliminar comentario (requiere autenticación)","operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/categorias/{id}":{"get":{"tags":["categoria-controller"],"operationId":"getCategoriaById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["categoria-controller"],"summary":"Actualizar categoría (requiere autenticación)","operationId":"updateCategoria","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Categoria"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["categoria-controller"],"summary":"Eliminar categoría (requiere autenticación)","operationId":"deleteCategoria","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/auth/usuarios/{id}":{"get":{"tags":["usuario-controller"],"operationId":"getUsuarioById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["usuario-controller"],"operationId":"updateUsuario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuario"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["usuario-controller"],"operationId":"deleteUsuario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/usuarios/favoritos":{"get":{"tags":["favorito-controller"],"operationId":"listFavoritos","parameters":[{"name":"idUsr","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["favorito-controller"],"summary":"Agregar favorito (requiere autenticación)","operationId":"addFavorito","parameters":[{"name":"idReceta","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["favorito-controller"],"summary":"Eliminar favorito (requiere autenticación)","operationId":"removeFavorito","parameters":[{"name":"idReceta","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"idUsr","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/stripe/webhook":{"post":{"tags":["stripe-webhook-controller"],"operationId":"handleWebhook","parameters":[{"name":"Stripe-Signature","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/recetas/":{"get":{"tags":["receta-controller"],"operationId":"getAllRecetas_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["receta-controller"],"summary":"Crear receta (requiere autenticación)","operationId":"createReceta","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Receta"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/paises/":{"get":{"tags":["pais-controller"],"operationId":"getAllPaises","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["pais-controller"],"summary":"Crear país (requiere autenticación)","operationId":"createPais","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Pais"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/megustas/":{"post":{"tags":["me-gusta-controller"],"summary":"Dar like (requiere autenticación)","operationId":"like","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/ingredientes/":{"get":{"tags":["ingrediente-controller"],"summary":"Listar ingredientes (público)","operationId":"listAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["ingrediente-controller"],"summary":"Crear ingrediente (requiere autenticación)","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ingrediente"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/estrellas/":{"post":{"tags":["estrella-controller"],"summary":"Agregar/valorar con estrella (requiere autenticación)","operationId":"rate","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/donaciones/create-session":{"post":{"tags":["donacion-controller"],"summary":"Crear sesión de donación (requiere autenticación)","operationId":"createSession","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/comentarios/":{"post":{"tags":["comentario-controller"],"summary":"Crear comentario (requiere autenticación)","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/categorias/":{"get":{"tags":["categoria-controller"],"operationId":"getAllCategorias_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["categoria-controller"],"summary":"Crear categoría (requiere autenticación)","operationId":"createCategoria","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Categoria"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/auth/usuarios":{"get":{"tags":["usuario-controller"],"operationId":"getAllUsuarios","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["usuario-controller"],"operationId":"createUsuario","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuario"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/login":{"post":{"tags":["usuario-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/admin/backup":{"post":{"tags":["admin-controller"],"summary":"Trigger backup (requiere autenticación)","operationId":"triggerBackup","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/recetas/{id}/ingredientes":{"get":{"tags":["receta-controller"],"operationId":"getIngredientesByReceta","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/recetas/trending":{"get":{"tags":["receta-controller"],"operationId":"getTrending","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":8}},{"name":"by","in":"query","required":false,"schema":{"type":"string","default":"visitas"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/recetas/pais/{idPais}":{"get":{"tags":["receta-controller"],"operationId":"getRecetasByPais","parameters":[{"name":"idPais","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/recetas/nombre/{nombre}":{"get":{"tags":["receta-controller"],"operationId":"getRecetaByNombre","parameters":[{"name":"nombre","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/recetas/del-dia":{"get":{"tags":["receta-controller"],"summary":"Receta del día","description":"Devuelve la receta seleccionada para el día actual. Estable por 24h.","operationId":"getRecetaDelDia","responses":{"200":{"description":"Receta encontrada","content":{"application/json":{}}},"404":{"description":"No hay recetas activas"}}}},"/recetas/categoria/{idCat}":{"get":{"tags":["receta-controller"],"operationId":"getRecetasByCategoria","parameters":[{"name":"idCat","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/recetas":{"get":{"tags":["receta-controller"],"operationId":"getAllRecetas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/paises/nombre/{nombre}":{"get":{"tags":["pais-controller"],"operationId":"getPaisesByNombre","parameters":[{"name":"nombre","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/megustas/count/{idReceta}":{"get":{"tags":["me-gusta-controller"],"operationId":"count","parameters":[{"name":"idReceta","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/ingredientes":{"get":{"tags":["ingrediente-controller"],"summary":"Listar ingredientes (público)","operationId":"listAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/estrellas/stats/{idReceta}":{"get":{"tags":["estrella-controller"],"operationId":"stats","parameters":[{"name":"idReceta","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/donaciones/{id}":{"get":{"tags":["donacion-controller"],"operationId":"getDonacion","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/comentarios/receta/{id}":{"get":{"tags":["comentario-controller"],"operationId":"getByReceta","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/categorias/nombre/{nombre}":{"get":{"tags":["categoria-controller"],"operationId":"getCategoriasByNombre","parameters":[{"name":"nombre","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/categorias":{"get":{"tags":["categoria-controller"],"operationId":"getAllCategorias","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/usuarios/":{"get":{"tags":["usuario-controller"],"operationId":"getAllUsuarios_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/perfil/":{"get":{"tags":["usuario-controller"],"operationId":"getPerfilAutenticado","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/perfil":{"get":{"tags":["usuario-controller"],"operationId":"getPerfilAutenticado_1","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/megustas":{"delete":{"tags":["me-gusta-controller"],"summary":"Eliminar like (requiere autenticación)","operationId":"unlike","parameters":[{"name":"idReceta","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"idUsr","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"Ingrediente":{"type":"object","properties":{"idIngrediente":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"receta":{"$ref":"#/components/schemas/Receta"}}},"Receta":{"type":"object","properties":{"idReceta":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"urlImagen":{"type":"string"},"ingredientes":{"type":"array","items":{"$ref":"#/components/schemas/Ingrediente"}},"preparacion":{"type":"string"},"estado":{"type":"integer","format":"int32"},"idCat":{"type":"integer","format":"int32"},"idPais":{"type":"integer","format":"int32"},"fechaCreacion":{"type":"string","format":"date"},"visitas":{"type":"integer","format":"int32"},"idUsr":{"type":"integer","format":"int32"}}},"Pais":{"type":"object","properties":{"idPais":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"urlImagen":{"type":"string"},"estado":{"type":"integer","format":"int32"},"fechaCreacion":{"type":"string","format":"date-time"},"comentario":{"type":"string"},"idUsr":{"type":"integer","format":"int32"}}},"Categoria":{"type":"object","properties":{"idCat":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"urlImagen":{"type":"string"},"estado":{"type":"integer","format":"int32"},"fechaCreacion":{"type":"string","format":"date-time"},"comentario":{"type":"string"},"idUsr":{"type":"integer","format":"int32"}}},"Perfil":{"type":"object","properties":{"idPerfil":{"type":"integer","format":"int32"},"nombre":{"type":"string"}}},"Usuario":{"type":"object","properties":{"idUsr":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"apellido":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"estado":{"type":"integer","format":"int32"},"fechaCreacion":{"type":"string","format":"date-time"},"comentario":{"type":"string"},"perfil":{"$ref":"#/components/schemas/Perfil"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}